+++ database/migrations/2019_05_03_134237_create_tests_section.php
@@ -21,7 +21,7 @@ class CreateTestsSection extends Migration
             $table->integer('unattempted')->nullable();
             $table->integer('correct')->nullable();
             $table->integer('incorrect')->nullable();
-            $table->integer('score')->nullable();
+            $table->float('score')->nullable();
             $table->bigInteger('time')->nullable();
             $table->timestamps();
         });
